redirect/requirements.txt
Flask
gunicorn

python-dotenv